refactor: eliminate duplicate types, unify constants, restore transaction feature
- Delete solana_transaction.rs (duplicate types) and solana_constants.rs (duplicate VKs)
- Reuse core types (Transaction, Action, ComplianceUnit, Delta) via feature gates
- Move LogicVerifierInputs from logic_proof.rs to logic_instance.rs (always available)
- Unify VK byte arrays in constants.rs as single source of truth
- Restore transaction feature to original meaning: ["compliance_circuit", "dep:sha3"]
- Add explicit k256 feature (dep:k256 doesn't create cfg)
- Gate k256/zkvm-dependent code behind respective features within modules
- Add Borsh derives via cfg_attr on Transaction, Action, ComplianceUnit, Delta, LogicVerifierInputs

Feature model:
  default = ["transaction", "prove", "k256"]
  solana  = ["transaction", "borsh", ...]  (no zkvm, no k256)
  zkvm    = ["dep:risc0-zkvm", "dep:lazy_static"]
