+++ arm/src/compliance_unit.rs
@@ -1,13 +1,18 @@
 //! Compliance unit module containing the compliance proof and instance.
 
+use crate::compliance::ComplianceInstance;
+#[cfg(any(feature = "zkvm", feature = "k256"))]
+use crate::error::ArmError;
+use serde::{Deserialize, Serialize};
+
+#[cfg(feature = "zkvm")]
 use crate::{
-    compliance::ComplianceInstance,
     constants::COMPLIANCE_VK,
-    error::ArmError,
     proving_system::{journal_to_instance, verify as verify_proof},
 };
+
+#[cfg(feature = "k256")]
 use k256::ProjectivePoint;
-use serde::{Deserialize, Serialize};
 
 #[cfg(feature = "prove")]
 use crate::{
@@ -19,6 +24,7 @@ use crate::{
 /// A compliance unit consists of a compliance proof and its corresponding instance.
 /// The vk is a constant in the compliance unit, so we don't place it here.
 #[derive(Clone, Debug, Deserialize, Serialize, PartialEq, Eq)]
+#[cfg_attr(feature = "borsh", derive(borsh::BorshSerialize, borsh::BorshDeserialize))]
 pub struct ComplianceUnit {
     /// The compliance proof (optional, would be absent when aggregation is enabled).
     pub proof: Option<Vec<u8>>,
@@ -26,11 +32,11 @@ pub struct ComplianceUnit {
     pub instance: ComplianceInstance,
 }
 
+#[cfg(feature = "prove")]
 impl ComplianceUnit {
     /// Creates a new compliance unit by proving the given witness.
     /// Proving key and verifying key are constants, so we don't need to pass
     /// them as parameters. Instance is generated by proving.
-    #[cfg(feature = "prove")]
     pub fn create(witness: &ComplianceWitness, proof_type: ProofType) -> Result<Self, ArmError> {
         let (proof, serialized_instance) = prove(COMPLIANCE_PK, witness, proof_type)?;
         let instance = journal_to_instance(&serialized_instance)?;
@@ -39,7 +45,10 @@ impl ComplianceUnit {
             instance,
         })
     }
+}
 
+#[cfg(feature = "zkvm")]
+impl ComplianceUnit {
     /// Verifies the compliance proof against the instance using the constant verifying key.
     pub fn verify(&self) -> Result<(), ArmError> {
         if let Some(proof) = &self.proof {
@@ -51,7 +60,10 @@ impl ComplianceUnit {
             ))
         }
     }
+}
 
+#[cfg(feature = "k256")]
+impl ComplianceUnit {
     /// Obtains the delta from the compliance instance.
     pub fn delta(&self) -> Result<ProjectivePoint, ArmError> {
         self.instance.delta_projective()
